@@ -1,28 +1,25 @@
 package d.d.meshenger;
 
 import android.Manifest;
-import android.app.Service;
 import android.content.BroadcastReceiver;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.content.ServiceConnection;
 import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.os.IBinder;
-import android.support.annotation.NonNull;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.content.LocalBroadcastManager;
-import android.support.v4.view.ViewPager;
-import android.support.v4.app.Fragment;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.widget.Toast;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+
+import androidx.annotation.NonNull;
+import androidx.core.app.ActivityCompat;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+import androidx.viewpager.widget.ViewPager;
+
+import com.google.android.material.tabs.TabLayout;
 
 import java.util.ArrayList;
 import java.util.Date;
@@ -30,18 +27,17 @@ import java.util.List;
 
 
 // the main view with tabs
-public class MainActivity extends MeshengerActivity implements ServiceConnection {
-    public MainService.MainBinder binder;
+public class MainActivity extends MeshengerActivity {
+    private static final String TAG = "MainActivity";
     private ViewPager mViewPager;
     private ContactListFragment contactListFragment;
     private EventListFragment eventListFragment;
     private SectionsPageAdapter sectionsPageAdapter;
     private int currentPage = 0;
-    private Date eventListAccessed;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        log("onCreate");
+        Log.d(TAG, "onCreate");
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
@@ -54,33 +50,8 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
         TabLayout tabLayout = findViewById(R.id.tabs);
         tabLayout.setupWithViewPager(mViewPager);
 
-        // ask for audio recording permissions
-        if (ContextCompat.checkSelfPermission(this, Manifest.permission.RECORD_AUDIO) != PackageManager.PERMISSION_GRANTED) {
-            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.RECORD_AUDIO}, 2);
-        }
-
-        this.eventListAccessed = new Date();
-
-        LocalBroadcastManager.getInstance(this).registerReceiver(refreshEventListReceiver, new IntentFilter("refresh_event_list"));
-        LocalBroadcastManager.getInstance(this).registerReceiver(refreshContactListReceiver, new IntentFilter("refresh_contact_list"));
-
-        bindService(new Intent(this, MainService.class), this, Service.BIND_AUTO_CREATE);
-    }
-
-    @Override
-    protected void onDestroy() {
-        log("onDestroy");
-        LocalBroadcastManager.getInstance(this).unregisterReceiver(refreshEventListReceiver);
-        LocalBroadcastManager.getInstance(this).unregisterReceiver(refreshContactListReceiver);
-
-        unbindService(this);
-        super.onDestroy();
-    }
-
-    @Override
-    public void onServiceConnected(ComponentName componentName, IBinder iBinder) {
-        log("OnServiceConnected");
-        this.binder = (MainService.MainBinder) iBinder;
+        LocalBroadcastManager.getInstance(this).registerReceiver(eventsChangedReceiver, new IntentFilter("events_changed"));
+        LocalBroadcastManager.getInstance(this).registerReceiver(contactsChangedReceiver, new IntentFilter("contacts_changed"));
 
         // in case the language has changed
         this.sectionsPageAdapter = new SectionsPageAdapter(getSupportFragmentManager());
@@ -91,39 +62,40 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
         mViewPager.addOnPageChangeListener(new ViewPager.OnPageChangeListener() {
             @Override
             public void onPageScrolled(int position, float positionOffset, int positionOffsetPixels) {
-               // ignore scrolling
+                // ignore scrolling
             }
 
             @Override
             public void onPageSelected(int position) {
-                log( "onPageSelected, position: " + position);
+                Log.d(TAG,  "onPageSelected, position: " + position);
                 MainActivity.this.currentPage = position;
                 if (position == 1) {
-                    MainActivity.this.eventListAccessed = new Date();
+                    MainService.instance.getEvents().setEventsViewedDate();
+                    //updateMissedCallsCounter();
                 }
             }
 
             @Override
             public void onPageScrollStateChanged(int state) {
-               // ignore scrolling
+                // ignore scrolling
             }
         });
 
         contactListFragment.refreshContactList();
         eventListFragment.refreshEventList();
-
-        this.binder.pingContacts();
     }
 
     @Override
-    public void onServiceDisconnected(ComponentName componentName) {
-        log("OnServiceDisconnected");
-        this.binder = null;
+    protected void onDestroy() {
+        super.onDestroy();
+        Log.d(TAG, "onDestroy");
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(eventsChangedReceiver);
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(contactsChangedReceiver);
     }
 
     @Override
     public boolean onOptionsItemSelected(MenuItem item) {
-        log("onOptionsItemSelected");
+        Log.d(TAG,"onOptionsItemSelected");
         int id = item.getItemId();
 
         switch (id) {
@@ -154,19 +126,10 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
         return super.onOptionsItemSelected(item);
     }
 
+    // update missed calls in events fragment title
     private void updateMissedCallsCounter() {
-        // update missed called notification
-        if (MainActivity.this.binder == null) {
-            return;
-        }
-
-        int missedCalls = 0;
-        Date since = this.eventListAccessed;
-        for (CallEvent event : this.binder.getEventsCopy()) {
-            if (event.date.compareTo(since) >= 0 && event.isMissedCall()) {
-                missedCalls += 1;
-            }
-        }
+        int missedCalls = MainService.instance.getEvents()
+            .getMissedCalls().size();
 
         this.sectionsPageAdapter.missedCalls = missedCalls;
         this.mViewPager.setAdapter(this.sectionsPageAdapter);
@@ -174,16 +137,16 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
         this.mViewPager.setCurrentItem(this.currentPage);
     }
 
-    private BroadcastReceiver refreshEventListReceiver = new BroadcastReceiver() {
+    private BroadcastReceiver eventsChangedReceiver = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
+            Log.d(TAG, "received events_changed");
             eventListFragment.refreshEventList();
-
             updateMissedCallsCounter();
         }
     };
 
-    private BroadcastReceiver refreshContactListReceiver = new BroadcastReceiver() {
+    private BroadcastReceiver contactsChangedReceiver = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
             contactListFragment.refreshContactList();
@@ -192,28 +155,59 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
 
     @Override
     protected void onResume() {
-        log("OnResume");
+        Log.d(TAG, "OnResume");
         super.onResume();
+
+        updateMissedCallsCounter();
+
+        checkPermissions(); // TODO: remove
     }
 
-    @Override
-    protected void onPause() {
-        log("onPause");
-        super.onPause();
+    // TODO: move to CallActivity
+    private void checkPermissions() {
+        Log.d(TAG, "checkPermissions");
+        Settings settings = MainService.instance.getSettings();
+        if (settings.getRecordAudio()) {
+            if (!Utils.hasPermission(this, Manifest.permission.RECORD_AUDIO)) {
+                ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.RECORD_AUDIO}, 1);
+                return;
+            }
+        }
+
+        if (settings.getRecordVideo()) {
+            if (!Utils.hasPermission(this, Manifest.permission.CAMERA)) {
+                ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.CAMERA}, 2);
+                return;
+            }
+        }
     }
 
     @Override
     public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
-        if (grantResults[0] != PackageManager.PERMISSION_GRANTED) {
-            Toast.makeText(this, R.string.permission_mic, Toast.LENGTH_LONG).show();
-            finish();
+        Settings settings = MainService.instance.getSettings();
+
+        switch (requestCode) {
+            case 1:
+                if (grantResults[0] != PackageManager.PERMISSION_GRANTED) {
+                    Toast.makeText(this, "Microphone disabled by default", Toast.LENGTH_LONG).show();
+                    settings.setRecordAudio(false);
+                }
+                break;
+            case 2:
+                if (grantResults[0] != PackageManager.PERMISSION_GRANTED) {
+                    Toast.makeText(this, "Camera disabled by default", Toast.LENGTH_LONG).show();
+                    settings.setRecordVideo(false);
+                }
+                break;
+            default:
+                Log.e(TAG, "Unknown permission requestCode: " + requestCode);
         }
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        log("onCreateOptionsMenu");
+        Log.d(TAG, "onCreateOptionsMenu");
         getMenuInflater().inflate(R.menu.menu_main_activity, menu);
         return true;
     }
@@ -234,7 +228,7 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
 
         @Override
         public CharSequence getPageTitle(int position) {
-            Log.d(this, "getPageTitle");
+            Log.d(TAG, "getPageTitle");
             if (mFragmentList.get(position) instanceof EventListFragment) {
                 if (this.missedCalls > 0) {
                     return mFragmentTitleList.get(position) + " (" + missedCalls + ")";
@@ -254,9 +248,5 @@ public class MainActivity extends MeshengerActivity implements ServiceConnection
         public int getCount() {
             return mFragmentList.size();
         }
-    }
-
-    private void log(String s) {
-        Log.d(this, s);
     }
 }
