@@ -1,18 +1,12 @@
 package d.d.meshenger;
 
-import android.app.Service;
-import android.content.ComponentName;
 import android.content.Intent;
-import android.content.ServiceConnection;
 import android.os.Bundle;
-import android.os.IBinder;
 import android.widget.TextView;
-import android.widget.Toast;
 
 
-public class AboutActivity extends MeshengerActivity implements ServiceConnection {
-    private int versionClicked = 0;
-    private MainService.MainBinder binder;
+public class AboutActivity extends MeshengerActivity {
+    private final String TAG = "AboutActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -29,38 +23,5 @@ public class AboutActivity extends MeshengerActivity implements ServiceConnectio
             Intent intent = new Intent(this, LicenseActivity.class);
             startActivity(intent);
         });
-
-        findViewById(R.id.versionTv).setOnClickListener(v -> {
-            if (binder != null) {
-                versionClicked += 1;
-                if (versionClicked < 4) {
-                    Toast.makeText(this, (4 - versionClicked) + " Clicks left for Development Mode", Toast.LENGTH_SHORT).show();
-                } else {
-                    if (!binder.getSettings().getDevelopmentMode()) {
-                        binder.getSettings().setDevelopmentMode(true);
-                        binder.saveDatabase();
-                    }
-                    Toast.makeText(this, "Development Mode Active", Toast.LENGTH_SHORT).show();
-                }
-            }
-        });
-
-        bindService(new Intent(this, MainService.class), this, Service.BIND_AUTO_CREATE);
-    }
-
-    @Override
-    protected void onDestroy() {
-        unbindService(this);
-        super.onDestroy();
-    }
-
-    @Override
-    public void onServiceConnected(ComponentName componentName, IBinder iBinder) {
-        this.binder = (MainService.MainBinder) iBinder;
-    }
-
-    @Override
-    public void onServiceDisconnected(ComponentName componentName) {
-        this.binder = null;
     }
 }
