@@ -1,12 +1,7 @@
 package d.d.meshenger;
 
 import android.app.Activity;
-import android.app.Service;
-import android.content.ComponentName;
-import android.content.Intent;
-import android.content.ServiceConnection;
 import android.graphics.Color;
-import android.os.IBinder;
 import android.os.Bundle;
 import android.text.Editable;
 import android.text.TextWatcher;
@@ -20,28 +15,29 @@ import android.widget.Spinner;
 import android.widget.TextView;
 import android.widget.Toast;
 
+import java.net.InetSocketAddress;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
 
-public class AddressActivity extends MeshengerActivity implements ServiceConnection {
-    private MainService.MainBinder binder;
-    Spinner storedAddressSpinner;
-    Spinner systemAddressSpinner;
-    Button pickStoredAddressButton;
-    Button pickSystemAddressButton;
-    EditText addressEditText;
-    Button addButton;
-    Button removeButton;
-    Button saveButton;
-    Button abortButton;
+public class AddressActivity extends MeshengerActivity {
+    private static final String TAG = "AddressActivity";
+    private Spinner storedAddressSpinner;
+    private Spinner systemAddressSpinner;
+    private Button pickStoredAddressButton;
+    private Button pickSystemAddressButton;
+    private EditText addressEditText;
+    private Button addButton;
+    private Button removeButton;
+    private Button saveButton;
+    private Button abortButton;
 
-    List<AddressEntry> systemAddressList;
-    List<AddressEntry> storedAddressList;
-    AddressListAdapter storedAddressListAdapter;
-    AddressListAdapter systemAddressListAdapter;
+    private List<AddressEntry> systemAddressList;
+    private List<AddressEntry> storedAddressList;
+    private AddressListAdapter storedAddressListAdapter;
+    private AddressListAdapter systemAddressListAdapter;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -60,7 +56,7 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
         saveButton = findViewById(R.id.SaveButton);
         abortButton = findViewById(R.id.AbortButton);
 
-        systemAddressList = Utils.collectAddresses();
+        systemAddressList = AddressUtils.getOwnAddresses();
         storedAddressList = new ArrayList<>();
         storedAddressListAdapter = new AddressListAdapter(this, Color.parseColor("#39b300")); //dark green
         systemAddressListAdapter = new AddressListAdapter(this, Color.parseColor("#b3b7b2")); //light grey
@@ -143,8 +139,8 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
             for (AddressEntry ae : this.storedAddressList) {
                 addresses.add(ae.address);
             }
-            this.binder.getSettings().setAddresses(addresses);
-            this.binder.saveDatabase();
+            MainService.instance.getSettings().setAddresses(addresses);
+            MainService.instance.saveDatabase();
             Toast.makeText(this, R.string.done, Toast.LENGTH_SHORT).show();
         });
 
@@ -152,36 +148,12 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
             finish();
         });
 
-        bindService();
-    }
-
-    @Override
-    protected void onDestroy() {
-        super.onDestroy();
-        unbindService(this);
-    }
-
-    private void bindService() {
-        // ask MainService to get us the binder object
-        Intent serviceIntent = new Intent(this, MainService.class);
-        bindService(serviceIntent, this, Service.BIND_AUTO_CREATE);
-    }
-
-    @Override
-    public void onServiceConnected(ComponentName componentName, IBinder iBinder) {
-        this.binder = (MainService.MainBinder) iBinder;
-
         // get from settings
-        for (String address : this.binder.getSettings().getAddresses()) {
+        for (String address : MainService.instance.getSettings().getAddresses()) {
             this.storedAddressList.add(parseAddress(address));
         }
 
         updateSpinners();
-    }
-
-    @Override
-    public void onServiceDisconnected(ComponentName componentName) {
-        this.binder = null;
     }
 
     public class AddressListAdapter extends BaseAdapter {
@@ -315,7 +287,7 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
      * Create AddressEntry from address string.
      * Do not perform any domain lookup
     */
-    AddressEntry parseAddress(String address) {
+    private AddressEntry parseAddress(String address) {
         // instead of parsing, lookup in known addresses first
         AddressEntry ae = AddressEntry.findAddressEntry(systemAddressList, address);
         if (ae != null) {
@@ -329,7 +301,7 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
             // IP address
             boolean mc = false;
             try {
-                mc = Utils.parseInetSocketAddress(address, 0).getAddress().isMulticastAddress();
+                mc = InetSocketAddress.createUnresolved(address, 0).getAddress().isMulticastAddress();
             } catch (Exception e) {
                 // ignore
             }
@@ -338,14 +310,5 @@ public class AddressActivity extends MeshengerActivity implements ServiceConnect
             // domain
             return new AddressEntry(address, "", false);
         }
-    }
-
-    @Override
-    protected void onResume() {
-        super.onResume();
-    }
-
-    private void log(String s) {
-        Log.d(this, s);
     }
 }
