@@ -4,8 +4,6 @@ import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.Canvas;
 import android.graphics.Paint;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -13,10 +11,14 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import java.util.List;
 
 
 class ContactListAdapter extends ArrayAdapter<Contact> {
+    private static final String TAG = "ContactListAdapter";
     private List<Contact> contacts;
     private Context context;
 
@@ -41,18 +43,46 @@ class ContactListAdapter extends ArrayAdapter<Contact> {
 
         ((TextView) convertView.findViewById(R.id.contact_name)).setText(contact.getName());
 
-        if (contact.getState() != Contact.State.PENDING) {
-            convertView.findViewById(R.id.contact_waiting).setVisibility(View.GONE);
-            ImageView state = convertView.findViewById(R.id.contact_state);
-            state.setVisibility(View.VISIBLE);
+        //if (contact.getState() != Contact.State.UNKNOWN) {
+            //convertView.findViewById(R.id.contact_waiting).setVisibility(View.GONE); // no animation
+            ImageView stateView = convertView.findViewById(R.id.contact_state);
+            stateView.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    contact.setState(Contact.State.UNKNOWN);
+                    // TODO: send ping
+                }
+            });
+
+            //stateView.setVisibility(View.VISIBLE);
             Bitmap bitmap = Bitmap.createBitmap(200, 200, Bitmap.Config.ARGB_8888);
             Canvas canvas = new Canvas(bitmap);
             Paint p = new Paint();
 
-            if (contact.getState() == Contact.State.ONLINE) {
-                p.setColor(0xFF7AE12D); // green
-            } else {
-                p.setColor(0xFFEC3E3E); // red
+            switch (contact.getState()) {
+                case ONLINE:
+                    p.setColor(0xFF7AE12D); // green
+                    float pc = (System.currentTimeMillis() - contact.getStateLastUpdated()) / Contact.STATE_TIMEOUT;
+                    if (pc >= 0.0 && pc <= 1.0) {
+                        p.setColor(0xFFFCBA03); // orange
+                        //canvas.drawCircle(100, 100, 70, p);
+                        canvas.drawArc(
+                        100, // left 
+                        100, // top
+                        100, // right 
+                        100, // bottom 
+                        0, // startAngle
+                        360 * pc, // sweepAngle
+                        false, // useCenter
+                        p);
+                    }
+                    break;
+                case OFFLINE:
+                    p.setColor(0xFFEC3E3E); // red
+                    break;
+                default:
+                    p.setColor(0xFFB5B5B5); // gray
+                    break;
             }
 
             canvas.drawCircle(100, 100, 100, p);
@@ -63,8 +93,8 @@ class ContactListAdapter extends ArrayAdapter<Contact> {
                 canvas.drawCircle(100, 100, 70, p);
             }
 
-            state.setImageBitmap(bitmap);
-        }
+            stateView.setImageBitmap(bitmap);
+        //}
 /*
         if (contact.recent) {
             contact.recent = false;
@@ -74,9 +104,5 @@ class ContactListAdapter extends ArrayAdapter<Contact> {
         }
 */
         return convertView;
-    }
-
-    private void log(String s) {
-        Log.d(this, s);
     }
 }
