@@ -5,10 +5,6 @@ import android.content.Intent;
 import android.content.res.Resources;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
@@ -20,15 +16,26 @@ import android.widget.AdapterView;
 import android.widget.EditText;
 import android.widget.ListView;
 import android.widget.PopupMenu;
+import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.fragment.app.Fragment;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import org.json.JSONException;
-
 import java.util.List;
 
 import static android.os.Looper.getMainLooper;
 
+import d.d.meshenger.call.CallActivity;
+import d.d.meshenger.call.DirectRTCClient;
+
 
 public class ContactListFragment extends Fragment implements AdapterView.OnItemClickListener {
+    private static final String TAG = "ContactListFragment";
     private ListView contactListView;
     private boolean fabExpanded = false;
     private FloatingActionButton fabScan;
@@ -39,7 +46,7 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        log("onCreateView");
+        Log.d(TAG, "onCreateView");
         View view = inflater.inflate(R.layout.fragment_contact_list, container, false);
 
         mainActivity = (MainActivity) getActivity();
@@ -58,7 +65,7 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
     }
 
     private void runFabAnimation(View fab) {
-        log("runFabAnimation");
+        Log.d(TAG, "runFabAnimation");
         AnimationSet scanSet = new AnimationSet(this.mainActivity, null);
         AnimationSet generateSet = new AnimationSet(this.mainActivity, null);
 
@@ -123,7 +130,8 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
         builder.setCancelable(false); // prevent key shortcut to cancel dialog
 
         builder.setPositiveButton(R.string.yes, (DialogInterface dialog, int id) -> {
-            this.mainActivity.binder.deleteContact(publicKey);
+            MainService.instance.getContacts().deleteContact(publicKey);
+            refreshContactListBroadcast();
             dialog.cancel();
         });
 
@@ -137,14 +145,14 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
     }
 
     void refreshContactList() {
-        log("refreshContactList");
-        if (this.mainActivity == null || this.mainActivity.binder == null) {
-            log("refreshContactList early return");
+        Log.d(TAG, "refreshContactList");
+        if (this.mainActivity == null) {
+            Log.d(TAG, "refreshContactList early return");
             return;
         }
 
         new Handler(getMainLooper()).post(() -> {
-            List<Contact> contacts = ContactListFragment.this.mainActivity.binder.getContactsCopy();
+            List<Contact> contacts = MainService.instance.getContacts().getContactListCopy();
             contactListView.setAdapter(new ContactListAdapter(ContactListFragment.this.mainActivity, R.layout.item_contact, contacts));
             contactListView.setOnItemClickListener(ContactListFragment.this);
             contactListView.setOnItemLongClickListener((AdapterView<?> adapterView, View view, int i, long l) -> {
@@ -156,11 +164,13 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
                 String block = res.getString(R.string.block);
                 String unblock = res.getString(R.string.unblock);
                 String share = res.getString(R.string.share);
+                String ping = "Ping"; //res.getString(R.string.ping);
                 String qr = "QR-ify";
 
                 menu.getMenu().add(delete);
                 menu.getMenu().add(rename);
                 menu.getMenu().add(share);
+                menu.getMenu().add(ping);
                 if (contact.getBlocked()) {
                     menu.getMenu().add(unblock);
                 } else {
@@ -181,9 +191,12 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
                         setBlocked(publicKey, true);
                     } else if (title.equals(unblock)) {
                         setBlocked(publicKey, false);
+                    } else if (title.equals(ping)) {
+                    	// TODO: ping contact
+                    	Log.d(TAG, "Ping not implemented here");
                     } else if (title.equals(qr)) {
                         Intent intent = new Intent(ContactListFragment.this.mainActivity, QRShowActivity.class);
-                        intent.putExtra("EXTRA_CONTACT", contact);
+                        intent.putExtra("EXTRA_CONTACT", contact.getPublicKey());
                         startActivity(intent);
                     }
                     return false;
@@ -194,20 +207,26 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
         });
     }
 
+    private void refreshContactListBroadcast() {
+        LocalBroadcastManager.getInstance(getActivity()).sendBroadcast(new Intent("contact_changed"));
+    }
+
     private void setBlocked(byte[] publicKey, boolean blocked) {
-        Contact contact = this.mainActivity.binder.getContactByPublicKey(publicKey);
+        Contacts contacts = MainService.instance.getContacts();
+        Contact contact = contacts.getContactByPublicKey(publicKey);
         if (contact != null) {
             contact.setBlocked(blocked);
-            this.mainActivity.binder.addContact(contact);
+            contacts.addContact(contact);
+            refreshContactListBroadcast();
         }
     }
 
-    private void shareContact(Contact c) {
-        log("shareContact");
+    private void shareContact(Contact contact) {
+        Log.d(TAG, "shareContact");
         try {
             Intent intent = new Intent(Intent.ACTION_SEND);
             intent.setType("text/plain");
-            intent.putExtra(Intent.EXTRA_TEXT, Contact.exportJSON(c, false).toString());
+            intent.putExtra(Intent.EXTRA_TEXT, Contact.toJSON(contact).toString());
             startActivity(intent);
         } catch (JSONException e) {
             e.printStackTrace();
@@ -215,7 +234,9 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
     }
 
     private void showContactEditDialog(byte[] publicKey, String name) {
-        log("showContactEditDialog");
+        Log.d(TAG, "showContactEditDialog");
+
+        Contact contact = MainService.instance.getContacts().getContactByPublicKey(publicKey);
         EditText et = new EditText(this.mainActivity);
         et.setText(name);
         AlertDialog dialog = new AlertDialog.Builder(this.mainActivity)
@@ -223,14 +244,26 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
             .setView(et)
             .setNegativeButton(getResources().getString(R.string.cancel), null)
             .setPositiveButton(R.string.ok, (DialogInterface dialogInterface, int i) -> {
-                String newName = et.getText().toString();
-                if (newName.length() > 0) {
-                    Contact contact = this.mainActivity.binder.getContactByPublicKey(publicKey);
-                    if (contact != null) {
-                        contact.setName(newName);
-                        this.mainActivity.binder.addContact(contact);
-                    }
+                String newName = et.getText().toString().trim();
+                if (newName.equals(contact.getName())) {
+                    // nothing to do
+                    return;
                 }
+
+                if (!Utils.isValidContactName(newName)) {
+                    Toast.makeText(getContext(), "Invalid name.", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+
+                if (null != MainService.instance.getContacts().getContactByName(newName)) {
+                    Toast.makeText(getContext(), "A contact with that name already exists.", Toast.LENGTH_SHORT).show();
+                    return;
+                }
+
+                // rename contact
+                contact.setName(newName);
+                MainService.instance.saveDatabase();
+                refreshContactListBroadcast();
             }).show();
     }
 
@@ -242,15 +275,12 @@ public class ContactListFragment extends Fragment implements AdapterView.OnItemC
 
     @Override
     public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-        log("onItemClick");
-        Contact contact = this.mainActivity.binder.getContactsCopy().get(i);
-        Intent intent = new Intent(this.mainActivity, CallActivity.class);
-        intent.setAction("ACTION_OUTGOING_CALL");
-        intent.putExtra("EXTRA_CONTACT", contact);
-        startActivity(intent);
-    }
-
-    private static void log(String s) {
-        Log.d(ContactListFragment.class.getSimpleName(), s);
+        Log.d(TAG, "onItemClick");
+        Contact contact = (Contact) adapterView.getAdapter().getItem(i);
+        if (DirectRTCClient.createOutgoingCall(contact)) {
+            Intent intent = new Intent(getContext(), CallActivity.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            startActivity(intent);
+        }
     }
 }
