@@ -8,71 +8,59 @@ import java.io.IOException;
 import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.Iterator;
+import java.util.List;
 
 
 class Database {
-    Settings settings;
-    ArrayList<Contact> contacts;
-    static String version = "3.1.0"; // current version
+    private static final String TAG = "Database";
+    static String version = "4.0.0"; // current version
+    private Settings settings;
+    private Contacts contacts;
+    private Events events;
 
     Database() {
-        this.contacts = new ArrayList<>();
+        this.contacts = new Contacts();
         this.settings = new Settings();
+        this.events = new Events();
     }
 
-    public void addContact(Contact contact) {
-        int idx = findContact(contact.getPublicKey());
-        if (idx >= 0) {
-            // contact exists - replace
-            this.contacts.set(idx, contact);
-        } else {
-            this.contacts.add(contact);
-        }
+    public Settings getSettings() {
+        return settings;
     }
 
-    public void deleteContact(byte[] publicKey) {
-        int idx = this.findContact(publicKey);
-        if (idx >= 0) {
-            this.contacts.remove(idx);
-        }
+    public Contacts getContacts() {
+        return contacts;
     }
 
-    public int findContact(byte[] publicKey) {
-        for (int i = 0; i < this.contacts.size(); i += 1) {
-            if (Arrays.equals(this.contacts.get(i).getPublicKey(), publicKey)) {
-                return i;
-            }
-        }
-        return -1;
+    public Events getEvents() {
+        return events;
     }
 
     public void onDestroy() {
         // zero keys from memory
-        if (this.settings.getSecretKey() != null) {
-            Arrays.fill(this.settings.getSecretKey(), (byte) 0);
+        if (settings.getSecretKey() != null) {
+            Arrays.fill(settings.getSecretKey(), (byte) 0);
         }
 
-        if (this.settings.getPublicKey() != null) {
-            Arrays.fill(this.settings.getPublicKey(), (byte) 0);
+        if (settings.getPublicKey() != null) {
+            Arrays.fill(settings.getPublicKey(), (byte) 0);
         }
 
-        for (Contact contact : this.contacts) {
+        for (Contact contact : contacts.getContactList()) {
             if (contact.getPublicKey() != null) {
                 Arrays.fill(contact.getPublicKey(), (byte) 0);
             }
         }
     }
 
-    public static Database load(String path, String password) throws IOException, JSONException {
-        // read database file
-        byte[] data = Utils.readExternalFile(path);
-
+    public static Database fromData(byte[] data, String password) throws IOException, JSONException {
         // encrypt database
         if (password != null && password.length() > 0) {
             data = Crypto.decryptDatabase(data, password.getBytes());
 
             if (data == null) {
-                throw new IOException("wrong database password.");
+                throw new IOException("Wrong database password.");
             }
         }
 
@@ -80,18 +68,16 @@ class Database {
             new String(data, Charset.forName("UTF-8"))
         );
 
-        boolean upgraded = upgradeDatabase(obj.getString("version"), Database.version, obj);
+        upgradeDatabase(obj.getString("version"), Database.version, obj);
         Database db = Database.fromJSON(obj);
 
-        if (upgraded) {
-            log("store updated database");
-            Database.store(path, db, password);
-        }
+        Log.d(TAG, "Loaded " + db.contacts.getContactList().size() + " contacts");
+        Log.d(TAG, "Loaded " + db.events.getEventList().size() + " events.");
 
         return db;
     }
 
-    public static void store(String path, Database db, String password) throws IOException, JSONException {
+    public static byte[] toData(Database db, String password) throws IOException, JSONException {
         JSONObject obj = Database.toJSON(db);
         byte[] data = obj.toString().getBytes();
 
@@ -100,21 +86,57 @@ class Database {
             data = Crypto.encryptDatabase(data, password.getBytes());
         }
 
-        // write database file
-        Utils.writeExternalFile(path, data);
+        Log.d(TAG, "Stored " + db.contacts.getContactList().size() + " contacts");
+        Log.d(TAG, "Stored " + db.events.getEventList().size() + " events.");
+
+        return data;
     }
 
-    private static boolean upgradeDatabase(String from, String to, JSONObject obj) throws JSONException {
+    private static void alignSettings(JSONObject settings) throws JSONException {
+        JSONObject defaults = Settings.toJSON(new Settings());
+
+        // default keys
+        Iterator<String> defaults_iter = defaults.keys();
+        List<String> defaults_keys = new ArrayList<>();
+        while (defaults_iter.hasNext()) {
+            defaults_keys.add(defaults_iter.next());
+        }
+
+        // current keys
+        Iterator<String> settings_iter = settings.keys();
+        List<String> settings_keys = new ArrayList<>();
+        while (settings_iter.hasNext()) {
+            settings_keys.add(settings_iter.next());
+        }
+
+        // add missing keys
+        for (String key : defaults_keys) {
+            if (!settings.has(key)) {
+                settings.put(key, defaults.get(key));
+            }
+        }
+
+        // remove extra keys
+        for (String key : settings_keys) {
+            if (!defaults.has(key)) {
+                settings.remove(key);
+            }
+        }
+    }
+
+    private static boolean upgradeDatabase(String from, String to, JSONObject db) throws JSONException {
         if (from.equals(to)) {
             return false;
         }
 
-        log("upgrade database from " + from + " to " + to);
+        Log.d(TAG, "Upgrade database from " + from + " to " + to);
+
+        JSONObject settings = db.getJSONObject("settings");
 
         // 2.0.0 => 2.1.0
         if (from.equals("2.0.0")) {
             // add blocked field (added in 2.1.0)
-            JSONArray contacts = obj.getJSONArray("contacts");
+            JSONArray contacts = db.getJSONArray("contacts");
             for (int i = 0; i < contacts.length(); i += 1) {
                 contacts.getJSONObject(i).put("blocked", false);
             }
@@ -124,8 +146,8 @@ class Database {
         // 2.1.0 => 3.0.0
         if (from.equals("2.1.0")) {
             // add new fields
-            obj.getJSONObject("settings").put("ice_servers", new JSONArray());
-            obj.getJSONObject("settings").put("development_mode", false);
+            settings.put("ice_servers", new JSONArray());
+            settings.put("development_mode", false);
             from = "3.0.0";
         }
 
@@ -138,7 +160,7 @@ class Database {
         // 3.0.1 => 3.0.2
         if (from.equals("3.0.1")) {
             // fix typo in setting name
-            obj.getJSONObject("settings").put("night_mode", obj.getJSONObject("settings").getBoolean("might_mode"));
+            settings.put("night_mode", settings.getBoolean("might_mode"));
             from = "3.0.2";
         }
 
@@ -148,13 +170,23 @@ class Database {
             from = "3.0.3";
         }
 
-        // 3.0.3 => 3.1.0
+        // 3.0.3 => 4.0.0
         if (from.equals("3.0.3")) {
-            obj.getJSONObject("settings").remove("language");
-            from = "3.1.0";
+            from = "4.0.0";
+            JSONObject events = new JSONObject();
+            events.put("entries", new JSONArray());
+            settings.put("events", events);
+
+            JSONObject contacts = new JSONObject();
+            JSONArray entries = db.getJSONArray("contacts");
+            contacts.put("entries", entries);
+            settings.put("contacts", contacts);
         }
 
-        obj.put("version", from);
+        // add missing keys with defaults and remove unexpected keys
+        alignSettings(settings);
+
+        db.put("version", from);
 
         return true;
     }
@@ -162,14 +194,9 @@ class Database {
     public static JSONObject toJSON(Database db) throws JSONException {
         JSONObject obj = new JSONObject();
         obj.put("version", db.version);
-        obj.put("settings", Settings.exportJSON(db.settings));
-
-        JSONArray contacts = new JSONArray();
-        for (Contact contact : db.contacts) {
-            contacts.put(Contact.exportJSON(contact, true));
-        }
-        obj.put("contacts", contacts);
-
+        obj.put("settings", Settings.toJSON(db.settings));
+        obj.put("contacts", Contacts.toJSON(db.contacts));
+        obj.put("events", Events.toJSON(db.events));
         return obj;
     }
 
@@ -180,21 +207,17 @@ class Database {
         db.version = obj.getString("version");
 
         // import contacts
-        JSONArray array = obj.getJSONArray("contacts");
-        for (int i = 0; i < array.length(); i += 1) {
-            db.contacts.add(
-                Contact.importJSON(array.getJSONObject(i), true)
-            );
-        }
+        JSONObject contacts = obj.getJSONObject("contacts");
+        db.contacts = Contacts.fromJSON(contacts);
 
         // import settings
         JSONObject settings = obj.getJSONObject("settings");
-        db.settings = Settings.importJSON(settings);
+        db.settings = Settings.fromJSON(settings);
+
+        // import events
+        JSONObject events = obj.getJSONObject("events");
+        db.events = Events.fromJSON(events);
 
         return db;
-    }
-
-    private static void log(String s) {
-        Log.d("Database", s);
     }
 }
