@@ -1,8 +1,6 @@
 package d.d.meshenger;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.text.format.DateUtils;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -11,18 +9,24 @@ import android.widget.ArrayAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+
 import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.List;
 
+import static d.d.meshenger.call.DirectRTCClient.CallDirection.INCOMING;
+import static d.d.meshenger.call.DirectRTCClient.CallDirection.OUTGOING;
 
-class EventListAdapter extends ArrayAdapter<CallEvent> {
+
+class EventListAdapter extends ArrayAdapter<Event> {
     private LayoutInflater inflater;
-    private List<CallEvent> events;
+    private List<Event> events;
     private List<Contact> contacts;
     private Context context;
 
-    public EventListAdapter(@NonNull Context context, int resource, List<CallEvent> events, List<Contact> contacts) {
+    public EventListAdapter(@NonNull Context context, int resource, List<Event> events, List<Contact> contacts) {
         super(context, resource, events);
 
         this.events = events;
@@ -32,9 +36,17 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
         inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
     }
 
-    public void update(List<CallEvent> events, List<Contact> contacts) {
+    public void update(List<Event> events, List<Contact> contacts) {
         this.events = events;
         this.contacts = contacts;
+    }
+
+    private String formatAddress(String address) {
+        if (address.startsWith("/")) {
+            return address.substring(1);
+        } else {
+            return address;
+        }
     }
 
     @Override
@@ -43,7 +55,7 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
     }
 
     @Override
-    public CallEvent getItem(int position) {
+    public Event getItem(int position) {
         return events.get(position);
     }
 
@@ -56,7 +68,7 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
     @Override
     public View getView(int position, @Nullable View view, @NonNull ViewGroup parent) {
         // show list in reverse, latest element first
-        CallEvent event = this.events.get(this.events.size() - position - 1);
+        Event event = this.events.get(this.events.size() - position - 1);
 
         if (view == null) {
             view = inflater.inflate(R.layout.item_event, null);
@@ -65,7 +77,7 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
         // find name
         String name = "";
         for (Contact contact : this.contacts) {
-            if (Arrays.equals(contact.getPublicKey(), event.pubKey)) {
+            if (Arrays.equals(contact.getPublicKey(), event.publicKey)) {
                 name = contact.getName();
                 break;
             }
@@ -73,7 +85,8 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
 
         TextView name_tv = view.findViewById(R.id.call_name);
         if (name.isEmpty()) {
-            name_tv.setText(this.context.getResources().getString(R.string.unknown_caller));
+            String unknown_caller = Utils.getUnknownCallerName(this.context, event.publicKey);
+            name_tv.setText(unknown_caller);
         } else {
             name_tv.setText(name);
         }
@@ -83,42 +96,52 @@ class EventListAdapter extends ArrayAdapter<CallEvent> {
             SimpleDateFormat ft = new SimpleDateFormat("'Today at' hh:mm:ss");
             date_tv.setText(ft.format(event.date));
         } else {
-            SimpleDateFormat ft = new SimpleDateFormat("yyyy.MM.dd 'at' hh:mm:ss");
+            SimpleDateFormat ft = new SimpleDateFormat("yyyy.MM.dd 'at' hh:mm");
             date_tv.setText(ft.format(event.date));
         }
 
         ImageView type_iv = view.findViewById(R.id.call_type);
-        switch (event.type) {
-            case INCOMING_ACCEPTED:
-            case INCOMING_DECLINED:
-                type_iv.setImageResource(R.drawable.call_incoming);
-                break;
-            case INCOMING_UNKNOWN:
-            case INCOMING_MISSED:
-            case INCOMING_ERROR:
-                type_iv.setImageResource(R.drawable.call_incoming_missed);
-                break;
-            case OUTGOING_ACCEPTED:
-            case OUTGOING_DECLINED:
-                type_iv.setImageResource(R.drawable.call_outgoing);
-                break;
-            case OUTGOING_UNKNOWN:
-            case OUTGOING_MISSED:
-            case OUTGOING_ERROR:
-                type_iv.setImageResource(R.drawable.call_outgoing_missed);
-                break;
+
+        if (event.callDirection == INCOMING) {
+            switch (event.callType) {
+                case ACCEPTED:
+                    type_iv.setImageResource(R.drawable.ic_incoming_call_accepted);
+                    break;
+                case DECLINED:
+                    type_iv.setImageResource(R.drawable.ic_incoming_call_declined);
+                    break;
+                case MISSED:
+                    type_iv.setImageResource(R.drawable.ic_incoming_call_missed);
+                    break;
+                case ERROR:
+                    type_iv.setImageResource(R.drawable.ic_incoming_call_error);
+                    break;
+            }
+        }
+
+        if (event.callDirection == OUTGOING) {
+            switch (event.callType) {
+                case ACCEPTED:
+                    type_iv.setImageResource(R.drawable.ic_outgoing_call_accepted);
+                    break;
+                case DECLINED:
+                    type_iv.setImageResource(R.drawable.ic_outgoing_call_declined);
+                    break;
+                case MISSED:
+                    type_iv.setImageResource(R.drawable.ic_outgoing_call_missed);
+                    break;
+                case ERROR:
+                    type_iv.setImageResource(R.drawable.ic_outgoing_call_error);
+                    break;
+            }
         }
 
         TextView address_tv = view.findViewById(R.id.call_address);
         if (event.address != null) {
-            address_tv.setText("(" + event.address.getHostAddress() + ")");
+            address_tv.setText("(" + formatAddress(event.address) + ")");
         } else {
             address_tv.setText("");
         }
         return view;
-    }
-
-    private static void log(String s) {
-        Log.d(EventListAdapter.class.getSimpleName(), s);
     }
 }
