@@ -1,14 +1,11 @@
 package d.d.meshenger;
 
+import android.app.Activity;
 import android.app.Dialog;
 import android.content.Intent;
 import android.content.res.Resources;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.Nullable;
-import android.support.design.widget.FloatingActionButton;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.PopupMenu;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
@@ -17,17 +14,30 @@ import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ListView;
+import android.widget.PopupMenu;
 import android.widget.Toast;
 
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+
+import java.net.Inet6Address;
+import java.net.InetAddress;
+import java.net.InetSocketAddress;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
+
+import d.d.meshenger.call.CallActivity;
+import d.d.meshenger.call.DirectRTCClient;
 
 import static android.os.Looper.getMainLooper;
 
 
 public class EventListFragment extends Fragment implements AdapterView.OnItemClickListener {
-    private MainActivity mainActivity;
+    private static final String TAG = "EventListFragment";
     private ListView eventListView;
     private EventListAdapter eventListAdapter;
     private FloatingActionButton fabDelete;
@@ -36,16 +46,15 @@ public class EventListFragment extends Fragment implements AdapterView.OnItemCli
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.fragment_event_list, container, false);
-        mainActivity = (MainActivity) getActivity();
 
         eventListView = view.findViewById(R.id.eventList);
         fabDelete = view.findViewById(R.id.fabDelete);
         fabDelete.setOnClickListener(v -> {
-            mainActivity.binder.clearEvents();
+            MainService.instance.getEvents().clear();
             refreshEventList();
         });
 
-        eventListAdapter = new EventListAdapter(mainActivity, R.layout.item_event, Collections.emptyList(), Collections.emptyList());
+        eventListAdapter = new EventListAdapter(getActivity(), R.layout.item_event, Collections.emptyList(), Collections.emptyList());
         eventListView.setAdapter(eventListAdapter);
         eventListView.setOnItemClickListener(this);
 
@@ -54,31 +63,30 @@ public class EventListFragment extends Fragment implements AdapterView.OnItemCli
         return view;
     }
 
-    void refreshEventList() {
-        log("refreshEventList");
-
-        if (this.mainActivity == null || this.mainActivity.binder == null) {
-            log("refreshEventList early return");
-            return;
-        }
-
-        List<CallEvent> events = EventListFragment.this.mainActivity.binder.getEventsCopy();
-        List<Contact> contacts = EventListFragment.this.mainActivity.binder.getContactsCopy();
+    public void refreshEventList() {
+        Log.d(TAG, "refreshEventList");
 
         new Handler(getMainLooper()).post(() -> {
-            log("refreshEventList update: " + events.size());
+            if (eventListAdapter == null) {
+                return;
+            }
+
+            List<Event> events = MainService.instance.getEvents().getEventListCopy();
+            List<Contact> contacts = MainService.instance.getContacts().getContactListCopy();
+
+            Log.d(TAG, "refreshEventList update: " + events.size());
             eventListAdapter.update(events, contacts);
             eventListAdapter.notifyDataSetChanged();
             eventListView.setAdapter(eventListAdapter);
 
             eventListView.setOnItemLongClickListener((AdapterView<?> adapterView, View view, int i, long l) -> {
-                CallEvent event = events.get(i);
-                PopupMenu menu = new PopupMenu(EventListFragment.this.mainActivity, view);
+                Event event = events.get(i);
+                PopupMenu menu = new PopupMenu(EventListFragment.this.getActivity(), view);
                 Resources res = getResources();
                 String add = res.getString(R.string.add);
                 String block = res.getString(R.string.block);
                 String unblock = res.getString(R.string.unblock);
-                Contact contact = EventListFragment.this.mainActivity.binder.getContactByPublicKey(event.pubKey);
+                Contact contact = MainService.instance.getContacts().getContactByPublicKey(event.publicKey);
 
                 // allow to add unknown caller
                 if (contact == null) {
@@ -112,45 +120,56 @@ public class EventListFragment extends Fragment implements AdapterView.OnItemCli
         });
     }
 
-    private void setBlocked(CallEvent event, boolean blocked) {
-        Contact contact = EventListFragment.this.mainActivity.binder.getContactByPublicKey(event.pubKey);
+    private void setBlocked(Event event, boolean blocked) {
+        Contact contact = MainService.instance.getContacts().getContactByPublicKey(event.publicKey);
         if (contact != null) {
             contact.setBlocked(blocked);
-            mainActivity.binder.saveDatabase();
+            MainService.instance.saveDatabase();
         } else {
-            // unknown contact
+            Log.w(TAG, "Cannot block: no contact found for public key");
         }
     }
 
-    private void showAddDialog(CallEvent event) {
-        log("showAddDialog");
+    private void showAddDialog(Event event) {
+        Log.d(TAG, "showAddDialog");
 
-        Dialog dialog = new Dialog(this.mainActivity);
+        Activity activity = getActivity();
+        Dialog dialog = new Dialog(activity);
         dialog.setContentView(R.layout.dialog_add_contact);
 
-        EditText nameEditText = dialog.findViewById(R.id.nameEditText);
-        Button exitButton = dialog.findViewById(R.id.ExitButton);
+        EditText nameEditText = dialog.findViewById(R.id.NameEditText);
+        Button cancelButton = dialog.findViewById(R.id.CancelButton);
         Button okButton = dialog.findViewById(R.id.OkButton);
 
         okButton.setOnClickListener((View v) -> {
             String name = nameEditText.getText().toString();
+            List<String> addresses = new ArrayList<>();
+
+            if (event.publicKey == null) {
+                Toast.makeText(activity, "Public key not set.", Toast.LENGTH_SHORT).show();
+                return;
+            }
 
             if (name.isEmpty()) {
-                Toast.makeText(this.mainActivity, R.string.contact_name_empty, Toast.LENGTH_SHORT).show();
+                Toast.makeText(activity, R.string.contact_name_empty, Toast.LENGTH_SHORT).show();
                 return;
             }
 
-            if (EventListFragment.this.mainActivity.binder.getContactByName(name) != null) {
-                Toast.makeText(this.mainActivity, R.string.contact_name_exists, Toast.LENGTH_LONG).show();
+            if (MainService.instance.getContacts().getContactByName(name) != null) {
+                Toast.makeText(activity, R.string.contact_name_exists, Toast.LENGTH_LONG).show();
                 return;
             }
 
-            String address = Utils.getGeneralizedAddress(event.address);
-            EventListFragment.this.mainActivity.binder.addContact(
-                new Contact(name, event.pubKey, Arrays.asList(address))
+            if (event.address != null) {
+                String address = getGeneralizedAddress(event.address);
+                addresses.add(address);
+            }
+
+            MainService.instance.getContacts().addContact(
+                new Contact(name, event.publicKey, addresses, false)
             );
 
-            Toast.makeText(this.mainActivity, R.string.done, Toast.LENGTH_SHORT).show();
+            Toast.makeText(activity, R.string.done, Toast.LENGTH_SHORT).show();
 
             refreshEventList();
 
@@ -158,29 +177,51 @@ public class EventListFragment extends Fragment implements AdapterView.OnItemCli
             dialog.dismiss();
         });
 
-        exitButton.setOnClickListener((View v) -> {
+        cancelButton.setOnClickListener((View v) -> {
             dialog.dismiss();
         });
 
         dialog.show();
     }
 
-    @Override
-    public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
-        log("onItemClick");
-        CallEvent event = this.eventListAdapter.getItem(i);
+    /*
+     * When adding an unknown contact, try to
+     * extract a MAC address from the IPv6 address.
+     */
+    private static String getGeneralizedAddress(String address) {
+        InetAddress addr = InetSocketAddress.createUnresolved(address, 0).getAddress();
+        if (addr instanceof Inet6Address) {
+            // if the IPv6 address contains a MAC address, take that.
+            byte[] mac = AddressUtils.getEUI64MAC((Inet6Address) addr);
+            if (mac != null) {
+                return Utils.bytesToMacAddress(mac);
+            }
+        }
 
-        String address = Utils.getGeneralizedAddress(event.address);
-        Contact contact = new Contact("", event.pubKey, Arrays.asList(address));
-        contact.setLastWorkingAddress(Utils.parseInetSocketAddress(address, MainService.serverPort));
-
-        Intent intent = new Intent(this.mainActivity, CallActivity.class);
-        intent.setAction("ACTION_OUTGOING_CALL");
-        intent.putExtra("EXTRA_CONTACT", contact);
-        startActivity(intent);
+        return address;
     }
 
-    private static void log(String s) {
-        Log.d(EventListFragment.class.getSimpleName(), s);
+    @Override
+    public void onItemClick(AdapterView<?> adapterView, View view, int i, long l) {
+        Log.d(TAG, "onItemClick");
+        Event event = this.eventListAdapter.getItem(i);
+
+        if (event.address == null) {
+            Toast.makeText(this.getContext(), "No address set for call!", Toast.LENGTH_SHORT).show();
+            return;
+        }
+
+        Contact contact = MainService.instance.getContacts().getContactByPublicKey(event.publicKey);
+        if (contact == null) {
+            String address = getGeneralizedAddress(event.address);
+            contact = new Contact("", event.publicKey, Arrays.asList(address), false);
+            contact.addAddress(address);
+        }
+
+        if (DirectRTCClient.createOutgoingCall(contact)) {
+            Intent intent = new Intent(getContext(), CallActivity.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+            startActivity(intent);
+        }
     }
 }
