@@ -5,84 +5,136 @@ import android.app.NotificationChannel;
 import android.app.NotificationManager;
 import android.app.PendingIntent;
 import android.app.Service;
+import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.content.IntentFilter;
 import android.graphics.BitmapFactory;
 import android.graphics.Color;
 import android.os.Binder;
 import android.os.Build;
 import android.os.Handler;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
-import android.support.v4.app.NotificationCompat;
-import android.support.v4.content.ContextCompat;
-import android.support.v4.content.LocalBroadcastManager;
+import android.text.format.DateUtils;
 import android.widget.Toast;
 
-import org.json.JSONObject;
-import org.libsodium.jni.Sodium;
+import androidx.annotation.Nullable;
+import androidx.core.app.NotificationCompat;
+import androidx.core.content.ContextCompat;
+import androidx.localbroadcastmanager.content.LocalBroadcastManager;
 
 import java.io.File;
 import java.io.IOException;
-import java.net.InetSocketAddress;
 import java.net.ServerSocket;
 import java.net.Socket;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
+import java.util.Timer;
+import java.util.TimerTask;
 
-import static android.support.v4.app.NotificationCompat.PRIORITY_MIN;
+import d.d.meshenger.call.CallActivity;
+import d.d.meshenger.call.DirectRTCClient;
 
 
 public class MainService extends Service implements Runnable {
-    private Database db = null;
+    private static final String TAG = "MainService";
+    private static DirectRTCClient currentCall = null;
+    private static Object currentCallLock = new Object();
+    private Database database = null;
     private boolean first_start = false;
     private String database_path = "";
     private String database_password = "";
+    private ServerSocket server;
+    private volatile boolean run = true;
+
+    public static MainService instance = null;
+    private final IBinder mBinder = new LocalBinder();
+    private NotificationCompat.Builder notificationBuilder;
 
     public static final int serverPort = 10001;
-    private ServerSocket server;
+    private static final int NOTIFICATION_ID = 42;
 
-    private volatile boolean run = true;
-    private RTCCall currentCall = null;
-
-    private ArrayList<CallEvent> events = null;
-
-    private MainBinder mainBinder = new MainBinder(this);
-
-    private int NOTIFICATION = 42;
+    /**
+     * Used to check whether the bound activity has really gone away and not unbound as part of an
+     * orientation change. We create a foreground service notification only if the former takes
+     * place.
+     */
+    private boolean mChangingConfiguration = false;
 
     @Override
     public void onCreate() {
         super.onCreate();
+        Log.d(TAG, "onCreate");
 
-        this.database_path = this.getFilesDir() + "/database.bin";
+        this.instance = this;
+        database_path = this.getFilesDir() + "/database.bin";
+
+        LocalBroadcastManager.getInstance(this).registerReceiver(eventsChangedReceiver, new IntentFilter("events_changed"));
+        LocalBroadcastManager.getInstance(this).registerReceiver(contactsChangedReceiver, new IntentFilter("contacts_changed"));
 
         // handle incoming connections
         new Thread(this).start();
-
-        events = new ArrayList<>();
     }
 
-    private void loadDatabase() {
+    public void loadDatabase() {
         try {
-            if ((new File(this.database_path)).exists()) {
+            if ((new File(database_path)).exists()) {
                 // open existing database
-                this.db = Database.load(this.database_path, this.database_password);
-                this.first_start = false;
+                byte[] data = Utils.readInternalFile(database_path);
+                database = Database.fromData(data, database_password);
+                first_start = false;
             } else {
                 // create new database
-                this.db = new Database();
-                this.first_start = true;
+                database = new Database();
+                first_start = true;
             }
         } catch (Exception e) {
-            // ignore
+            e.printStackTrace();
         }
     }
 
-    private void saveDatabase() {
+    public void replaceDatabase(Database database) {
+        if (database != null) {
+            if (this.database == null) {
+                this.database = database;
+            } else {
+                this.database = database;
+                saveDatabase();
+            }
+        }
+    }
+
+    public boolean isFirstStart() {
+        return first_start;
+    }
+
+    public String getDatabasePassword() {
+        return database_password;
+    }
+
+    public void setDatabasePassword(String password) {
+        database_password = password;
+    }
+
+    public Database getDatabase() {
+        return database;
+    }
+
+    public Settings getSettings() {
+        return database.getSettings();
+    }
+
+    public Contacts getContacts() {
+        return database.getContacts();
+    }
+
+    public Events getEvents() {
+        return database.getEvents();
+    }
+
+    public void saveDatabase() {
         try {
-            Database.store(MainService.this.database_path, MainService.this.db, MainService.this.database_password);
+            byte[] data = Database.toData(database, database_password);
+            Utils.writeInternalFile(database_path, data);
         } catch (Exception e) {
             e.printStackTrace();
         }
@@ -91,66 +143,35 @@ public class MainService extends Service implements Runnable {
     @Override
     public void onDestroy() {
         super.onDestroy();
-        this.run = false;
+        run = false;
+
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(eventsChangedReceiver);
+        LocalBroadcastManager.getInstance(this).unregisterReceiver(contactsChangedReceiver);
 
         // The database might be null here if no correct
         // database password was supplied to open it.
 
-        if (this.db != null) {
+        if (database != null) {
             try {
-                Database.store(this.database_path, this.db, this.database_password);
+                byte[] data = Database.toData(database, database_password);
+                Utils.writeInternalFile(database_path, data);
             } catch (Exception e) {
                 e.printStackTrace();
             }
         }
 
         // shutdown listening socket and say goodbye
-        if (this.db != null && this.server != null && this.server.isBound() && !this.server.isClosed()) {
+        if (database != null && server != null && server.isBound() && !server.isClosed()) {
             try {
-                byte[] ownPublicKey = this.db.settings.getPublicKey();
-                byte[] ownSecretKey = this.db.settings.getSecretKey();
-                String message = "{\"action\": \"status_change\", \"status\", \"offline\"}";
-
-                for (Contact contact : this.db.contacts) {
-                    if (contact.getState() == Contact.State.OFFLINE) {
-                        continue;
-                    }
-
-                    byte[] encrypted = Crypto.encryptMessage(message, contact.getPublicKey(), ownPublicKey, ownSecretKey);
-                    if (encrypted == null) {
-                        continue;
-                    }
-
-                    Socket socket = null;
-                    try {
-                        socket = contact.createSocket();
-                        if (socket == null) {
-                            continue;
-                        }
-
-                        PacketWriter pw = new PacketWriter(socket);
-                        pw.writeMessage(encrypted);
-                        socket.close();
-                    } catch (Exception e) {
-                        if (socket != null) {
-                            try {
-                                socket.close();
-                            } catch (Exception ee) {
-                                // ignore
-                            }
-                        }
-                    }
-                }
-
                 server.close();
             } catch (Exception e) {
                 e.printStackTrace();
             }
         }
 
-        if (this.db != null) {
+        if (database != null) {
             // zero keys from memory
-            this.db.onDestroy();
+            database.onDestroy();
         }
     }
 
@@ -169,17 +190,75 @@ public class MainService extends Service implements Runnable {
         PendingIntent pendingNotificationIntent = PendingIntent.getActivity(this, 0, notificationIntent, 0);
 
         Context mActivity = getApplicationContext();
-        Notification notification = new NotificationCompat.Builder(mActivity, channelId)
-                .setOngoing(true)
+        notificationBuilder = new NotificationCompat.Builder(mActivity, channelId)
+                //.setOngoing(true)
+                .setOnlyAlertOnce(true) // keep notification update from turning on the screen
                 .setSmallIcon(R.drawable.ic_logo)
                 .setLargeIcon(BitmapFactory.decodeResource(getResources(), R.drawable.logo_small))
-                .setPriority(PRIORITY_MIN)
+                .setPriority(NotificationCompat.PRIORITY_DEFAULT)
                 .setCategory(Notification.CATEGORY_SERVICE)
                 .setContentText(getResources().getText(R.string.listen_for_incoming_calls))
-                .setContentIntent(pendingNotificationIntent)
-                .build();
+                .setContentIntent(pendingNotificationIntent);
+                //.setVisibility(VISIBILITY_PUBLIC);
 
-        startForeground(NOTIFICATION, notification);
+        startForeground(NOTIFICATION_ID, notificationBuilder.build());
+
+
+        // update notification, since we want do display an updated "xxx ago"
+        Timer t = new Timer();
+        TimerTask tt = new TimerTask() {
+            @Override
+            public void run() {
+                Log.d(TAG, "trigger notification update");
+                updateMissedCallsNotification();
+            };
+        };
+
+        t.scheduleAtFixedRate(tt, 1000, 10 * 1000); // every 10s
+    }
+
+    private final BroadcastReceiver eventsChangedReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            Log.d(TAG, "received events_changed");
+            updateMissedCallsNotification();
+            saveDatabase();
+        }
+    };
+
+    private final BroadcastReceiver contactsChangedReceiver = new BroadcastReceiver() {
+        @Override
+        public void onReceive(Context context, Intent intent) {
+            Log.d(TAG, "received contacts_changed");
+            saveDatabase();
+        }
+    };
+
+    private void updateMissedCallsNotification() {
+        if (notificationBuilder == null || database == null) {
+            return;
+        }
+
+        NotificationManager manager = (NotificationManager) getSystemService(Context.NOTIFICATION_SERVICE);
+        List<Event> missedCalls = getEvents().getMissedCalls();
+
+        if (missedCalls.size() == 0) {
+            notificationBuilder.setContentText(
+                getResources().getText(R.string.listen_for_incoming_calls)
+            );
+        } else {
+            Event lastCall = missedCalls.get(missedCalls.size() - 1);
+            notificationBuilder.setContentText(
+                String.valueOf(missedCalls.size())
+                + " missed calls"
+                + " - "
+                + DateUtils.getRelativeTimeSpanString(
+                    lastCall.date.getTime(), System.currentTimeMillis(), 0L, DateUtils.FORMAT_ABBREV_ALL
+                )
+            );
+        }
+
+        manager.notify(NOTIFICATION_ID, notificationBuilder.build());
     }
 
     final static String START_FOREGROUND_ACTION = "START_FOREGROUND_ACTION";
@@ -194,7 +273,7 @@ public class MainService extends Service implements Runnable {
     public static void stop(Context ctx) {
         Intent stopIntent = new Intent(ctx, MainService.class);
         stopIntent.setAction(STOP_FOREGROUND_ACTION);
-        ctx.startService(stopIntent);
+        ContextCompat.startForegroundService(ctx, stopIntent);
     }
 
     @Override
@@ -202,147 +281,23 @@ public class MainService extends Service implements Runnable {
         if (intent == null || intent.getAction() == null) {
             // ignore
         } else if (intent.getAction().equals(START_FOREGROUND_ACTION)) {
-            log("Received Start Foreground Intent");
+            Log.d(TAG, "Received Start Foreground Intent");
             showNotification();
         } else if (intent.getAction().equals(STOP_FOREGROUND_ACTION)) {
-            log("Received Stop Foreground Intent");
-            ((NotificationManager) getSystemService(NOTIFICATION_SERVICE)).cancel(NOTIFICATION);
+            Log.d(TAG, "Received Stop Foreground Intent");
+            ((NotificationManager) getSystemService(NOTIFICATION_SERVICE)).cancel(NOTIFICATION_ID);
             stopForeground(true);
             stopSelf();
         }
         return START_NOT_STICKY;
     }
 
-    private void handleClient(MainBinder binder, Socket client) {
-        // just a precaution
-        if (this.db == null) {
-            return;
-        }
-
-        byte[] clientPublicKey = new byte[Sodium.crypto_sign_publickeybytes()];
-        byte[] ownSecretKey = this.db.settings.getSecretKey();
-        byte[] ownPublicKey = this.db.settings.getPublicKey();
-
-        try {
-            PacketWriter pw = new PacketWriter(client);
-            PacketReader pr = new PacketReader(client);
-            Contact contact = null;
-
-            InetSocketAddress remote_address = (InetSocketAddress) client.getRemoteSocketAddress();
-            log("incoming connection from " + remote_address);
-
-            while (true) {
-                byte[] request = pr.readMessage();
-                if (request == null) {
-                    break;
-                }
-
-                String decrypted = Crypto.decryptMessage(request, clientPublicKey, ownPublicKey, ownSecretKey);
-                if (decrypted == null) {
-                    log("decryption failed");
-                    break;
-                }
-
-                if (contact == null) {
-                    for (Contact c : this.db.contacts) {
-                        if (Arrays.equals(c.getPublicKey(), clientPublicKey)) {
-                            contact = c;
-                        }
-                    }
-
-                    if (contact == null && this.db.settings.getBlockUnknown()) {
-                        if (this.currentCall != null) {
-                            log("block unknown contact => decline");
-                            this.currentCall.decline();
-                        }
-                        break;
-                    }
-
-                    if (contact != null && contact.getBlocked()) {
-                        if (this.currentCall != null) {
-                            log("blocked contact => decline");
-                            this.currentCall.decline();
-                        }
-                        break;
-                    }
-
-                    if (contact == null) {
-                        // unknown caller
-                        contact = new Contact("", clientPublicKey.clone(), new ArrayList<>());
-                    }
-                }
-
-                // suspicious change of identity during connection...
-                if (!Arrays.equals(contact.getPublicKey(), clientPublicKey)) {
-                    log("suspicious change of identity");
-                    continue;
-                }
-
-                // remember last good address (the outgoing port is random and not the server port)
-                contact.setLastWorkingAddress(
-                    new InetSocketAddress(remote_address.getAddress(), MainService.serverPort)
-                );
-
-                JSONObject obj = new JSONObject(decrypted);
-                String action = obj.optString("action", "");
-
-                switch (action) {
-                    case "call": {
-                        // someone calls us
-                        log("got call...");
-                        String offer = obj.getString("offer");
-                        this.currentCall = new RTCCall(this, binder, contact, client, offer);
-
-                        // respond that we accept the call
-
-                        byte[] encrypted = Crypto.encryptMessage("{\"action\":\"ringing\"}", contact.getPublicKey(), ownPublicKey, ownSecretKey);
-                        pw.writeMessage(encrypted);
-
-                        Intent intent = new Intent(this, CallActivity.class);
-                        intent.setAction("ACTION_INCOMING_CALL");
-                        intent.putExtra("EXTRA_CONTACT", contact);
-                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-                        startActivity(intent);
-                        return;
-                    }
-                    case "ping": {
-                        log("got ping...");
-                        // someone wants to know if we are online
-                        binder.setContactState(contact.getPublicKey(), Contact.State.ONLINE);
-                        byte[] encrypted = Crypto.encryptMessage("{\"action\":\"pong\"}", contact.getPublicKey(), ownPublicKey, ownSecretKey);
-                        pw.writeMessage(encrypted);
-                        break;
-                    }
-                    case "status_change": {
-                        if (obj.optString("status", "").equals("offline")) {
-                            binder.setContactState(contact.getPublicKey(), Contact.State.OFFLINE);
-                        } else {
-                            log("Received unknown status_change: " + obj.getString("status"));
-                        }
-                    }
-                }
-            }
-
-            log("call disconnected");
-            LocalBroadcastManager.getInstance(this).sendBroadcast(new Intent("call_declined"));
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            log("client disconnected (exception)");
-            if (this.currentCall != null) {
-                this.currentCall.decline();
-            }
-        }
-
-        // zero out key
-        Arrays.fill(clientPublicKey, (byte) 0);
-    }
-
+    // runs in a thread
     @Override
     public void run() {
         try {
             // wait until database is ready
-            while (this.db == null && this.run) {
+            while (this.database == null && this.run) {
                 try {
                     Thread.sleep(1000);
                 } catch (Exception e) {
@@ -351,17 +306,30 @@ public class MainService extends Service implements Runnable {
             }
 
             server = new ServerSocket(serverPort);
-            MainBinder binder = (MainBinder) onBind(null);
+            LocalBinder binder = (LocalBinder) onBind(null);
 
             while (this.run) {
+                Socket socket = null;
                 try {
-                    Socket socket = server.accept();
-                    new Thread(() -> handleClient(binder, socket)).start();
+                    socket = server.accept();
+                    if (DirectRTCClient.createIncomingCall(socket)) {
+                        Intent intent = new Intent(this, CallActivity.class);
+                        intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                        startActivity(intent);
+                    }
+
+                    Thread.sleep(50); // mitigate DDOS attack
                 } catch (IOException e) {
-                    // ignore
+                    if (socket != null) {
+                        try {
+                            socket.close();
+                        } catch (IOException _e) {
+                            // ignore
+                        }
+                    }
                 }
             }
-        } catch (IOException e) {
+        } catch (Exception e) {
             e.printStackTrace();
             new Handler(getMainLooper()).post(() -> Toast.makeText(this, e.getMessage(), Toast.LENGTH_LONG).show());
             stopSelf();
@@ -369,226 +337,20 @@ public class MainService extends Service implements Runnable {
         }
     }
 
-    /*
-    * Allows communication between MainService and other objects
-    */
-    static class MainBinder extends Binder {
-        private MainService service;
-
-        MainBinder(MainService service) {
-            this.service = service;
-        }
-
-        Context getContext() {
-            return this.service;
-        }
-
-        RTCCall getCurrentCall() {
-            return this.service.currentCall;
-        }
-
-        boolean isFirstStart() {
-            return this.service.first_start;
-        }
-
-        Contact getContactByPublicKey(byte[] pubKey) {
-            for (Contact contact : this.service.db.contacts) {
-                if (Arrays.equals(contact.getPublicKey(), pubKey)) {
-                    return contact;
-                }
-            }
-            return null;
-        }
-
-        Contact getContactByName(String name) {
-            for (Contact contact : this.service.db.contacts) {
-                if (contact.getName().equals(name)) {
-                    return contact;
-                }
-            }
-            return null;
-        }
-
-        void addContact(Contact contact) {
-            this.service.db.addContact(contact);
-            saveDatabase();
-            LocalBroadcastManager.getInstance(this.service).sendBroadcast(new Intent("refresh_contact_list"));
-        }
-
-        void deleteContact(byte[] pubKey) {
-            this.service.db.deleteContact(pubKey);
-            saveDatabase();
-            LocalBroadcastManager.getInstance(this.service).sendBroadcast(new Intent("refresh_contact_list"));
-        }
-
-        void setContactState(byte[] publicKey, Contact.State state) {
-            Contact contact = getContactByPublicKey(publicKey);
-            if (contact != null && contact.getState() != state) {
-                contact.setState(state);
-                LocalBroadcastManager.getInstance(this.service).sendBroadcast(new Intent("refresh_contact_list"));
-            }
-        }
-
-        String getDatabasePassword() {
-            return this.service.database_password;
-        }
-
-        void setDatabasePassword(String password) {
-            this.service.database_password = password;
-        }
-
-        Database getDatabase() {
-            return this.service.db;
-        }
-
-        void loadDatabase() {
-            this.service.loadDatabase();
-        }
-
-        void replaceDatabase(Database db) {
-            if (db != null) {
-                if (this.service.db == null) {
-                    this.service.db = db;
-                } else {
-                    this.service.db = db;
-                    saveDatabase();
-                }
-            }
-        }
-
-        void pingContacts() {
-            Log.d(this, "pingContacts");
-            new Thread(new PingRunnable(
-                this,
-                getContactsCopy(),
-                getSettings().getPublicKey(),
-                getSettings().getSecretKey())
-            ).start();
-        }
-
-        void saveDatabase() {
-            this.service.saveDatabase();
-        }
-
-        Settings getSettings() {
-            return this.service.db.settings;
-        }
-
-        // return a cloned list
-        List<Contact> getContactsCopy() {
-           return new ArrayList<>(this.service.db.contacts);
-        }
-
-        void addCallEvent(Contact contact, CallEvent.Type type) {
-            InetSocketAddress last_working = contact.getLastWorkingAddress();
-            this.service.events.add(new CallEvent(
-                contact.getPublicKey(),
-                    (last_working != null) ? last_working.getAddress() : null,
-                type
-            ));
-            LocalBroadcastManager.getInstance(this.service).sendBroadcast(new Intent("refresh_event_list"));
-        }
-
-        // return a cloned list
-        List<CallEvent> getEventsCopy() {
-            return new ArrayList<>(this.service.events);
-        }
-
-        void clearEvents() {
-            this.service.events.clear();
-            LocalBroadcastManager.getInstance(this.service).sendBroadcast(new Intent("refresh_event_list"));
-        }
-    }
-
-    static class PingRunnable implements Runnable {
-        private List<Contact> contacts;
-        byte[] ownPublicKey;
-        byte[] ownSecretKey;
-        MainBinder binder;
-
-        PingRunnable(MainBinder binder, List<Contact> contacts, byte[] ownPublicKey, byte[] ownSecretKey) {
-            this.binder = binder;
-            this.contacts = contacts;
-            this.ownPublicKey = ownPublicKey;
-            this.ownSecretKey = ownSecretKey;
-        }
-
-        @Override
-        public void run() {
-            for (Contact contact : contacts) {
-                Socket socket = null;
-                byte[] publicKey = contact.getPublicKey();
-
-                try {
-                    socket = contact.createSocket();
-                    if (socket == null) {
-                        this.binder.setContactState(publicKey, Contact.State.OFFLINE);
-                        continue;
-                    }
-
-                    PacketWriter pw = new PacketWriter(socket);
-                    PacketReader pr = new PacketReader(socket);
-
-                    log("send ping to " + contact.getName());
-
-                    byte[] encrypted = Crypto.encryptMessage("{\"action\":\"ping\"}", publicKey, ownPublicKey, ownSecretKey);
-                    if (encrypted == null) {
-                        socket.close();
-                        continue;
-                    }
-
-                    pw.writeMessage(encrypted);
-
-                    byte[] request = pr.readMessage();
-                    if (request == null) {
-                        socket.close();
-                        continue;
-                    }
-
-                    String decrypted = Crypto.decryptMessage(request, publicKey, ownPublicKey, ownSecretKey);
-                    if (decrypted == null) {
-                        log("decryption failed");
-                        socket.close();
-                        continue;
-                    }
-
-                    JSONObject obj = new JSONObject(decrypted);
-                    String action = obj.optString("action", "");
-                    if (action.equals("pong")) {
-                        log("got pong");
-                        this.binder.setContactState(publicKey, Contact.State.ONLINE);
-                    }
-
-                    socket.close();
-                } catch (Exception e) {
-                    this.binder.setContactState(publicKey, Contact.State.OFFLINE);
-                    if (socket != null) {
-                        try {
-                            socket.close();
-                        } catch (Exception ee) {
-                            // ignore
-                        }
-                    }
-                    e.printStackTrace();
-                }
-            }
-
-            log("send refresh_contact_list");
-            LocalBroadcastManager.getInstance(this.binder.getContext()).sendBroadcast(new Intent("refresh_contact_list"));
-        }
-
-        private void log(String data) {
-            Log.d(this, data);
-        }
-    }
-
     @Nullable
     @Override
     public IBinder onBind(Intent intent) {
-        return mainBinder;
+        Log.d(TAG, "onBind");
+        return mBinder;
     }
 
-    private void log(String data) {
-        Log.d(this, data);
+    /**
+     * Class used for the client Binder.  Since this service runs in the same process as its
+     * clients, we don't need to deal with IPC.
+     */
+    public class LocalBinder extends Binder {
+        MainService getService() {
+            return MainService.this;
+        }
     }
 }
